+++ libraries/MS5611/CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.4.2] - 2025-08-31
+- fix #42, add reference to STM32 version of library
+- add **float getAltitude(float airPressure = 1013.25)**
+- add derived MS5607 class
+- update readme.md
+- update GitHub actions
+- minor edits
+
 ## [0.4.1] - 2024-10-23
 - add **float getPressurePascal()** Pascal == SI-unit
 - add **uint8_t getAddress()** convenience function.